requirements.txt
flask>=3.0.0
requests>=2.32.0
icalendar>=6.0.0
psycopg2-binary>=2.9
gunicorn>=22.0
